Set up initial application
